Add tenacityteam/tenacity drama
Add tenacityteam/tenacity drama:

Tenacity, an Audacity fork, is an easy-to-use multi-track audio editor and recorder. A GitHub user named Cookiengineer was the leader and one of the developers of Tenacity. Some 4chan users doxxed and attempted to murder Cookiengineer in a coordinated attack campaign in July 2021. Cookiengineer quit after the harassment and stabbing from the 4chan users.

Primary sources and archives:

https://github.com/tenacityteam/tenacity/issues/99
https://web.archive.org/web/20221012221409/github.com/tenacityteam/tenacity/issues/99
https://archive.is/GhhKi

Journalism:

https://www.theregister.com/2021/07/07/tenacity_maintainer_quits_4chan_harassment
https://kiwifarms.net/threads/open-source-software-community.38130/post-9384746
https://news.ycombinator.com/item?id=27765210